requirements.txt
Django==4.0.3
djangorestframework==3.13.1
celery==5.2.3
redis==4.1.4
python-dotenv==0.19.2
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.0.0
django-seed==0.3.1
Faker==13.3.4
six==1.16.0
toposort==1.7
python-dateutil==2.8.2
